producer/pyproject.toml
[project]
name = "producer"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Patrick Paulino", email = "ppaulino@ritchiebrothers.com" }
]
dependencies = [
    "flask>=3.0.3",
    "pact-python>=2.2.0",
    "tomli>=2.0.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[tool.rye]
managed = true
virtual = true
dev-dependencies = [
    "pytest>=8.2.2",
]

[tool.rye.scripts]
dev = { cmd = "flask run --debug --port 8082 --host 0.0.0.0", env = { FLASK_APP = "./app.py" } }

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::PendingDeprecationWarning",
]

[tool.pylint."MESSAGES CONTROL"]
disable = [
    "missing-function-docstring",
    "missing-module-docstring",
]